Add caching of JSON objects and `TextPart` objects
Experimental.

The `Content` objects of the history are included in every
new request, which means they are serialized to JSON repeatedly.

This changes a number of objects to cache their JSON representation.
Some can probably be made lazy, if they're not sent often.

Also tries to avoid creating too many composite objects of a type
where there is a finite number of possible different instances
(pairs of enum values).

Finally caches `TextPart` objects based on their text.
This is speculative. It only makes a difference if the same
text is received repeatedly.